[feat] Add support for plotting `brainunit.Quantity` instances in matplotlib (#54)
* feat: Add support for plotting `brainunit.Quantity` instances in matplotlib

* fix: Skip test if matplotlib is not installed

* update

* feat: Add support for importing `_matplotlib_compat` module

This commit adds support for importing the `_matplotlib_compat` module in the `brainunit` package. The module is now imported using the wildcard import syntax, allowing easier access to its contents. This change improves the usability and flexibility of the `brainunit` package.

* Update docs

* feat: Add support for setting axis unit in matplotlib plots

This commit adds a new function `set_axis_unit` to the `_matplotlib_compat` module. This function allows users to set the scale of the specified axis to a target unit in matplotlib plots. It takes parameters such as the axis index, target unit, matplotlib axis object, and precision for adjusting the axis labels. The function also updates the axis label with the target unit.

The commit also includes a new test case `test_set_axis_unit` in the `_matplotlib_compat_test` module to verify the functionality of the `set_axis_unit` function.

* Update _matplotlib_compat.py

* Update docs

* Add consistency check

* Update _matplotlib_compat.py

* Try add more registry

* add matplotlib support

* update doc

* fix bug

---------

Co-authored-by: Chaoming Wang <adaduo@outlook.com>